bugs

+++ run_command.c
@@ -60,8 +60,16 @@ void bKill(char** args, int argn) {
 			}
       	}	
       	else {
-			if(kill(to_be_killed,SIGTERM) == -1) {
-	  			perror("Kill failed\n");
+			if (job->status = JOBSTOP) {
+				kill(to_be_killed,SIGCONT);
+				if(kill(to_be_killed,SIGTERM) == -1) {
+	  				perror("Kill failed\n");
+				}
+			}
+			else {
+				if(kill(to_be_killed,SIGTERM) == -1) {
+	  				perror("Kill failed\n");
+				}
 			}
       	}
     }
